[refactor] Get rid of settings.py.
[ghstack-poisoned]
