+++ misc/generate_ir_design_doc.py
@@ -5,8 +5,8 @@ from taichi.core import settings
 
 
 def extract_doc(doc_filename=None):
-    repo_dir = settings.get_repo_directory()
-    statements_fn = os.path.join(repo_dir, 'taichi/ir/statements.h')
+    statements_fn = os.path.join(os.path.dirname(__file__),
+                                 '../taichi/ir/statements.h')
     with open(statements_fn, 'r') as f:
         statements = f.readlines()
