+++ python/taichi/tools/video.py
@@ -1,6 +1,6 @@
 import os
 
-from taichi.core.settings import get_os_name
+from taichi.core import get_os_name
 from taichi.misc.image import imwrite
 
 FRAME_FN_TEMPLATE = '%06d.png'
